requirements.txt
Django>=4.2,<5.0
opencv-python-headless>=4.8
numpy>=1.26
Pillow>=10.0
openai>=1.50.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
whitenoise>=6.7.0
dj-database-url>=2.2.0
psycopg2-binary>=2.9.9
